Make import_po work without iterating it
import_po is currently a generator of warnings. If you're not interested
in warnings though, it will stop running when it first encounters one.
This is quite annoying so I've changed it so that it is just a regular
function that returns a list of warnings when it's all done.