Prevent crash when a homepage is created in a synchronised locale source
Fixes #245.

This just works around the crash by disabling the auto-creation of
homepages when a new homepage is created in a locale that is set in the
"sync from" field of another.

The problem is that Treebeard calls the post_save signal before it's set
the numchild attribute of the parent. This isn't a problem in most cases
as the translations are created underneath other parent pages. But for
homepages, we need the parent `numchild` to be up to date because
translations of homepages are created as siblings.

Without the correct numchild, Treebeard/Wagtail attempts to create those
siblings with the same path which results in a crash.