Pass allowed page types into page chooser
Fixes #267

Some page choosers have limited page types which could be defined either
by the foreign key being set to a specific page type or a list of
allowed page types on the PageChooserPanel/Block.

This adds support for finding this information from the panel/block and
passing it into the widget on the frontend.

While implementing this, I discovered an inconsistency with the way
we're extracting related page segments from streamfield. On the model,
these are extracted as overridable segments, but in streamfield they were
being extracted as a translatable related object segments. This PR fixes
that as well.