Use field_path for finding segment names
This swaps out the 'content path' for the new 'field path' for when
we're looking for the field name a segment was extracted from.

The issue with content paths is they don't contain any information about
the type of the block the segment was extracted from, so we must always
go back to the original page.

The problem with this is if the block was deleted from the original
page, this will raise an error. So, instead, we use the newly introduced
'field path', which allows us to navigate to the block type that was
used.